Create workflow for checking country package versions.
related to PolicyEngine/issues#378

In order to block deployments to the api v1 we need some way to wait for the current country package versions
to actually show up in the simulation api before actually deploying the service.

This workflow will check in a loop for the country package version s3 metadata files and fail if they don't show up
after a timeout.