Create tagger service
related to PolicyEngine/issues#378

This update adds a new service (not yet deployed) that can be used to tag a release of simulation api
based on the us/uk country pacakge version data.

Includes build tests and was tested locally on desktop as well.

Also includes a fix to the way metadata was being published (Was not getting stored as valid json)