Corretly capture the current revision in metadata.
Fixes #237

This fixes an issue where we were recording metadata associated with the previous revision of the simulation api
instead of the one we just deployed.

This was happening because the cloud_run_api_v2 module will resolve the currently live revision name instead of the one it's deployed.

Change to explicitly defining the new revision name and then referencing that instead.